+++ src/kusto_tool/__init__.py
@@ -1,5 +1,5 @@
 """A high-level API for querying Kusto."""
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 from .database import Cluster, KustoDatabase, cluster
 from .expression import TableExpr